Use int instead of int64 in offers parsing

+++ database/offers.sql
@@ -1,8 +1,8 @@
 CREATE TABLE IF NOT EXISTS offers (
-    offer_id int,
-    seller_id int,
-    price int,
+    offer_id bigint,
+    seller_id bigint,
+    price bigint,
     name text,
-    quantity int,
+    quantity bigint,
     UNIQUE(offer_id, seller_id)
 );
