+++ models/offer.go
@@ -2,8 +2,8 @@ package models
 
 type Offer struct {
 	SellerID int64  `json:"seller_id"`
-	OfferID  int64  `json:"offer_id"`
+	OfferID  int    `json:"offer_id"`
 	Name     string `json:"name"`
-	Price    int64  `json:"price"`
-	Quantity int64  `json:"quantity"`
+	Price    int    `json:"price"`
+	Quantity int    `json:"quantity"`
 }
